add the log function to all routes

+++ src/utils/Crud.js
@@ -2,6 +2,7 @@ const statusCodes = require("http-status-codes").StatusCodes;
 const appResponder = require("./appResponder");
 const AppError = require("./AppError");
 const Query = require("./Query");
+const { ChangeTypes, logChanges } = require("./logChanges.util");
 
 class CRUD {
   constructor(model) {
@@ -9,8 +10,15 @@ class CRUD {
   }
 
   // Create a new record
-  async create(body, res) {
+  async create(body, res, req) {
     const data = await this.model.create(body);
+
+    await logChanges(
+      this.model.tableName,
+      data.id,
+      ChangeTypes.create,
+      req.user
+    );
     appResponder(statusCodes.CREATED, data, res);
   }
 
@@ -44,28 +52,69 @@ class CRUD {
     appResponder(statusCodes.OK, data, res);
   }
 
-  // Update record by id
   async update(id, res, req) {
-    const [updatedCount] = await this.model.update(req.body, {
-      where: { id },
-    });
+    const existing = await this.model.findByPk(id);
+    if (!existing) {
+      throw new AppError("Invalid Id, no such resource in the database", 404);
+    }
+    const existingPlain = existing.get({ plain: true });
+
+    const fieldsChanged = {};
+    for (const key in req.body) {
+      const newVal = req.body[key];
+      const oldVal = existingPlain[key];
+
+      if (String(newVal) !== String(oldVal)) {
+        fieldsChanged[key] = { before: oldVal, after: newVal };
+      }
+    }
+
+    const [updatedCount] = await this.model.update(req.body, { where: { id } });
 
     if (updatedCount === 0) {
-      throw new AppError("Invalid Id, no such resource in the database", 404);
+      throw new AppError("Failed to update record", 500);
     }
 
     const data = await this.model.findByPk(id);
 
+    if (Object.keys(fieldsChanged).length > 0) {
+      await logChanges(
+        this.model.tableName,
+        id,
+        ChangeTypes.update,
+        req.user,
+        fieldsChanged
+      );
+    }
+
     appResponder(statusCodes.OK, data, res);
   }
 
-  // Delete record by id
-  async delete(id, res) {
-    const deletedCount = await this.model.destroy({ where: { id } });
+  async delete(id, res, req) {
+    const existing = await this.model.findByPk(id);
+    if (!existing) {
+      throw new AppError("Invalid Id, no such resource in the database", 404);
+    }
 
+    const existingPlain = existing.get({ plain: true });
+
+    const deletedCount = await this.model.destroy({ where: { id } });
     if (deletedCount === 0) {
       throw new AppError("Invalid Id, no such resource in the database", 404);
     }
+
+    const fieldsChanged = {};
+    for (const key in existingPlain) {
+      fieldsChanged[key] = { before: existingPlain[key] };
+    }
+
+    await logChanges(
+      this.model.tableName,
+      id,
+      ChangeTypes.delete,
+      req.user,
+      fieldsChanged
+    );
 
     appResponder(
       statusCodes.OK,
